Input clean up
Rewrote some input stuff, made look at and look in their own functions,
and made that work better, so that if a string contains both of them,
then it will ask for clarification, e.g. look in bath towel will ask
look In or
AT? So that's better. Also, the game will notify you when your input
does nothing, and when you enter a room, it will automatically look.
Also, it will tell you when you pick something up. Next I need to move a
lot of the main function stuff into different methods for starting the
game to enable quitting to the menu, to reselect a game, and stuff,
without having to exit and restart the executable.