Room Export Implementation
Made a pair of functions that should adequately address the need for a
single line way to output the map. I just need to write the importer
now. Something worth remembering is that the current implementation
doesn't backwards associate the parent node from the child node, in the
adjacent list, to avoid infinite recursion.